+++ kononov.pavel/lab1/build.sh
@@ -9,4 +9,4 @@ fi
 
 chmod 666 "$pid_log"
 
-gcc -Wall -Werror daemon_source.cpp -o daemon_lab_5 -lstdc++ -std=c++11
+gcc -Wall -Werror main.cpp config_reader.cpp file_deleter.cpp daemon.cpp -o daemon_lab_5 -lstdc++ -std=c++11
