+++ kononov.pavel/lab1/main.cpp
@@ -1,0 +1,11 @@
+#include "daemon.h"
+
+int main(int argc, char** argv)
+{
+	Daemon instance = Daemon::getInstance();
+
+	instance.init(argc, argv);
+	instance.exec();
+
+	return 0;
+}
