--- kononov.pavel/lab2/interfaces/message.h
@@ -1,20 +1,0 @@
-#ifndef OS_LABS_SPBPU_2019_MESSAGE_H
-#define OS_LABS_SPBPU_2019_MESSAGE_H
-
-enum class Status
-{
-    ALIVE,
-    DEAD
-};
-
-struct Message
-{
-    Status status;
-    int number;
-
-    Message(Status st = Status::ALIVE, int num = 0) : status(st), number(num)
-    {
-    }
-};
-
-#endif //OS_LABS_SPBPU_2019_MESSAGE_H
