--- kononov.pavel/lab2/utils/utils.h
@@ -1,10 +1,0 @@
-#ifndef WOLF_UTILS_H
-#define WOLF_UTILS_H
-
-#define SEMAPHORE_HOST_NAME "LAB2_HOST"
-#define SEMAPHORE_CLIENT_NAME "LAB2_CLIENT"
-const int TIMEOUT = 5;
-
-int GetRand(int limit);
-
-#endif //WOLF_UTILS_H
